register user and email in lib and components
