--- src/components/firebase.js
@@ -1,22 +1,0 @@
-
- //import { getAnalytics } from "firebase/analytics";
- //import { } from "firebase/app";
- import {initializeApp} from 'https://www.gstatic.com/firebasejs/9.9.3/firebase-app.js' 
-
- const firebaseConfig = {
-     apiKey: "AIzaSyDGmzguJLCabUpc92fAl0oBvRZYuyP8R_U",
-     authDomain: "comuniapp-4741b.firebaseapp.com",
-     projectId: "comuniapp-4741b",
-     storageBucket: "comuniapp-4741b.appspot.com",
-     messagingSenderId: "691655579330",
-     appId: "1:691655579330:web:aef13ca9a633870a19cd5b",
-     measurementId: "G-97LP0J3GWT"
- };
-
- const app = initializeApp(firebaseConfig);
- //const analytics = getAnalytics(app);
-
-
-export function register(email, password) {
-  
-}
