+++ src/lib/firebase.js
@@ -1,0 +1,13 @@
+import { initializeApp } from 'https://www.gstatic.com/firebasejs/9.9.3/firebase-app.js'
+
+const firebaseConfig = {
+  apiKey: "AIzaSyDGmzguJLCabUpc92fAl0oBvRZYuyP8R_U",
+  authDomain: "comuniapp-4741b.firebaseapp.com",
+  projectId: "comuniapp-4741b",
+  storageBucket: "comuniapp-4741b.appspot.com",
+  messagingSenderId: "691655579330",
+  appId: "1:691655579330:web:aef13ca9a633870a19cd5b",
+  measurementId: "G-97LP0J3GWT"
+};
+
+export const app = initializeApp(firebaseConfig);
