+++ src/style.css
@@ -9,6 +9,8 @@ body {
 
 .welcome {
     background-color: rgb(255, 166, 0);
+    display: flex;
+    flex-direction: column;
 }
 
 .welcome img {
